pyproject.toml
[project]
name = "claudebridge"
version = "0.1.0"
description = "OpenAI-compatible API bridge for Claude using OAuth tokens"
authors = [{name = "Ylan Allouche"}]
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"

dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "loguru>=0.7.3",
    "waitress>=3.0.0",
    "pyhtml-htmx @ git+https://github.com/YlanAllouche/pyhtml-htmx/",
    "pyhtml-cem @ git+https://github.com/YlanAllouche/pyhtml-cem",
]

[project.scripts]
claudebridge = "claudebridge.__main__:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = {find = {}}
include-package-data = true

[tool.setuptools.package-data]
claudebridge = ["static/**/*"]


